yet more cleanup and one more noise issue

+++ initialize_json.py
@@ -23,7 +23,7 @@ employer_json = {
 }
 
 
-def generate_population_value(datafile: str, inception: date, start_count: int, mu: float, sigma: int) -> str:
+def generate_population_value(datafile: str, inception: date, start_count: int, mu: float, sigma: float) -> str:
     s_datafile = f'"datafile":\"{datafile}\"'
     s_start = f'"start":\"{inception}\"'
     s_pop = f'"pop":\"{start_count}\"'
@@ -40,7 +40,7 @@ def compile_json(company_name: str,
                  schedule: Schedule = Schedule.QUARTERLY,
                  datafile: str = '',
                  mu: float = 0.0,
-                 sigma: int = 0) -> dict:
+                 sigma: float = 0) -> dict:
 
     employer_json['name'] = company_name
     employer_json['schedule'] = schedule
